Editing problems

+++ Sparkle/src/profiler/spk_time_consumption_metrics.cpp
@@ -11,11 +11,11 @@ namespace spk
 	{
 		spk::JSON::Object result;
 
-		result.addAttribute(L"Min").set<long>(min());
-		result.addAttribute(L"Max").set<long>(max());
-		result.addAttribute(L"Average").set<long>(average());
-		result.addAttribute(L"Total duration").set<long>(cardinal());
-		result.addAttribute(L"Cardinal").set<long>(cardinal());
+		result.addAttribute(L"Min").set<long>(static_cast<long>(min()));
+		result.addAttribute(L"Max").set<long>(static_cast<long>(max()));
+		result.addAttribute(L"Average").set<long>(static_cast<long>(average()));
+		result.addAttribute(L"Total duration").set<long>(static_cast<long>(cardinal()));
+		result.addAttribute(L"Cardinal").set<long>(static_cast<long>(cardinal()));
 
 		return (result);
 	}
